moved src/app/store/ to src/app/state/
